Refactor SalaryUnboundException package structure, update rounding mode in User class, and clean up imports in UserUseCaseImpl and UserUseCaseTest
