+++ domain/model/src/main/java/co/com/pragma/model/user/exceptions/SalaryUnboundException.java
@@ -1,7 +1,4 @@
-package co.com.pragma.model.user;
-
-import co.com.pragma.model.user.exceptions.ErrorMessage;
-import co.com.pragma.model.user.exceptions.UserException;
+package co.com.pragma.model.user.exceptions;
 
 public class SalaryUnboundException extends UserException {
     public SalaryUnboundException() {
