+++ infrastructure/driven-adapters/r2dbc-mysql/src/main/java/co/com/pragma/r2dbc/UserEntityRepository.java
@@ -3,7 +3,6 @@ package co.com.pragma.r2dbc;
 import co.com.pragma.r2dbc.entity.UserEntity;
 import org.springframework.data.repository.query.ReactiveQueryByExampleExecutor;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
-import reactor.core.publisher.Mono;
 
 public interface UserEntityRepository extends ReactiveCrudRepository<UserEntity, Integer>, ReactiveQueryByExampleExecutor<UserEntity> {
